apps/mapa/requirements.txt
django-appconf